Data processing functional. Need to fix the loss functions (dimensions don't match)

+++ src/image_data_processing.py
@@ -46,43 +46,45 @@ def download_and_prepare_flickr8k_dataset(path: str) -> None:
 
     dataset_path = f"{path}/flickr8k"
     # Download dataset
-    kaggle.api.dataset_download_files(dataset_identifier, path=dataset_path, unzip=True)
+    # Only download the dataset if it hasn't been downloaded yet
+    if not os.path.exists(dataset_path):
+        kaggle.api.dataset_download_files(dataset_identifier, path=dataset_path, unzip=True)
 
-    # Prepare directories for processed data
-    if not os.path.exists(f"{dataset_path}/train"):
-        os.makedirs(f"{dataset_path}/train")
-    if not os.path.exists(f"{dataset_path}/val"):
-        os.makedirs(f"{dataset_path}/val")
-    if not os.path.exists(f"{dataset_path}/test"):
-        os.makedirs(f"{dataset_path}/test")
+        # Prepare directories for processed data
+        if not os.path.exists(f"{dataset_path}/train"):
+            os.makedirs(f"{dataset_path}/train")
+        if not os.path.exists(f"{dataset_path}/val"):
+            os.makedirs(f"{dataset_path}/val")
+        if not os.path.exists(f"{dataset_path}/test"):
+            os.makedirs(f"{dataset_path}/test")
 
-    # Define resize transformation
-    transform = transforms.Resize((224, 224))
+        # Define resize transformation
+        transform = transforms.Resize((224, 224))
 
-    images_list = os.listdir(f"{dataset_path}/Images")
-    # Split into train and validation
-    # 80% train, 20% validation
-    test_images = images_list[int(len(images_list) * 0.8) :]
-    train_images = images_list[: int(len(images_list) * 0.8)]
-    # Of the train images, 80% will be used for training and 20% for validation
-    val_images = train_images[int(len(train_images) * 0.8) :]
-    train_images = train_images[: int(len(train_images) * 0.8)]
-    # Process and save images
-    list_splits = ["train", "val", "test"]
-    list_class_dirs = [train_images, val_images, test_images]
-    for i in range(len(list_splits)):
-        split = list_splits[i]
-        list_images = list_class_dirs[i]
-        # Adjust according to the actual Flickr8k structure on disk
-        images_path = f"{dataset_path}/Images"
+        images_list = os.listdir(f"{dataset_path}/Images")
+        # Split into train and validation
+        # 80% train, 20% validation
+        test_images = images_list[int(len(images_list) * 0.8) :]
+        train_images = images_list[: int(len(images_list) * 0.8)]
+        # Of the train images, 80% will be used for training and 20% for validation
+        val_images = train_images[int(len(train_images) * 0.8) :]
+        train_images = train_images[: int(len(train_images) * 0.8)]
+        # Process and save images
+        list_splits = ["train", "val", "test"]
+        list_class_dirs = [train_images, val_images, test_images]
+        for i in range(len(list_splits)):
+            split = list_splits[i]
+            list_images = list_class_dirs[i]
+            # Adjust according to the actual Flickr8k structure on disk
+            images_path = f"{dataset_path}/Images"
 
-        for image_file in list_images:
-            image_path = f"{images_path}/{image_file}"
-            image = Image.open(image_path).convert("RGB")
-            image = transform(image)
-            image.save(f"{dataset_path}/{split}/{image_file}")
+            for image_file in list_images:
+                image_path = f"{images_path}/{image_file}"
+                image = Image.open(image_path).convert("RGB")
+                image = transform(image)
+                image.save(f"{dataset_path}/{split}/{image_file}")
 
-    shutil.rmtree(f"{dataset_path}/Images")
+        shutil.rmtree(f"{dataset_path}/Images")
 
     print("Dataset processed and saved.")
